server/requirements.txt
fastapi==0.115.14
uvicorn==0.35.0
python-dotenv==1.1.1
pydantic==2.6.1
pydantic-settings==2.1.0
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
alembic==1.13.1
asyncpg==0.29.0